move all modal styling to one file
