--- src/components/Modals/CreateChannelModal/CreateChannelModal.module.css
@@ -1,36 +1,0 @@
-.usernameDisplay {
-    background: white;
-    color: black;
-    padding: 5px;
-    border-radius: 5px;
-    margin-right: 5px;
-    width: auto;
-}
-
-.usernameDisplayWrapper {
-    width: 100%;
-    display: flex;
-    flex-wrap: wrap;
-    margin-bottom: 15px;
-}
-
-.errorMsg {
-	color: red;
-}
-
-.newUserInput {
-    display: flex;
-    flex-direction: column;
-    margin-top: 20px;
-}
-
-.newUserDisplay {
-    position: relative;
-    display: flex;
-    margin: 10px 0;
-    flex-direction: column;
-}
-
-.inviteMembersDisplay {
-    margin-top: 10px;
-}
