+++ src/components/Modals/OrgSettingsModal/OrgSettingsModal.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { connect } from "react-redux";
 import { actions } from "../../../context";
+
 import CustomButton from '../../UI/CustomButton/CustomButton';
-import CustomForm from '../../UI/CustomForm/CustomForm';
 import CustomModal from '../../UI/CustomModal/CustomModal';
 import InviteMembersModal from "../InviteMembersModal/InviteMembersModal";
 import CanView from "../../CanView/CanView";
