--- src/components/Modals/PendingInvitationsModal/PendingInvitationsModal.module.css
@@ -1,5 +1,0 @@
-.invitationDisplay {
-    display: flex;
-    justify-content: space-evenly;
-    padding: 10px;
-}
