+++ src/components/UI/CustomButton/CustomButton.module.css
@@ -9,14 +9,15 @@
     color: rgb(216, 212, 207);
     border: none;
     border-radius: 4px;
-    padding: 0px 11px;
+    padding: 0px 30px;
+    align-self: center;
 }
 
 .customButton:focus , .enter:focus{
     outline:none;
 }
 
-.customButton:hover, .enter:hover {
+.customButton:hover {
     background-color: rgb(16, 106, 82);;
     box-shadow: rgba(0, 0, 0, 0.3) 0px 1px 4px
 }
@@ -38,6 +39,11 @@
 
 .enter {
     float: right;
+    background-color: rgb(78, 25, 84);
+}
+
+.enter:hover {
+    box-shadow: rgba(0, 0, 0, 0.3) 0px 1px 4px
 }
