--- src/components/UI/CustomForm/CustomForm.module.css
@@ -1,8 +1,0 @@
-.customForm {
-	background-color: rgb(26,29,33);
-    color: rgb(209,210,211);
-	display: flex;
-	flex-direction: column;
-	min-height: 70px;
-	width: 100%;
-}
