+++ src/index.css
@@ -95,6 +95,7 @@ body {
 	padding: 0px;
 	margin: 0px;
 	text-shadow: none;
+	font-size: 2rem;
 }
 h1 {
     font-size:1.5em !important;
